Correct timeOut handling in readBlocking/writeBlocking.
Currently, if the EventFlag wait() returns due to a spurious wake
or if a notification was received but there is a need to go back
to waiting due to insufficient data to read or space to write,
the timeOut parameter is not adjusted to account for the time
already spent waiting. This CL fixes the same.

Bug: 34290175
Test: FMQ unit tests

Change-Id: I71399b23d88d5f3fc91ca1a9c0cf9d3ea35e8290