Remove logspam for MQs w/o event flag word
Some message queues do not have an event flag word.
The MQ constructor code was attempting to create one
regardless. This was leading to syslog spamming
with error messages from 'mapGrantorDesc' function.

The check inside 'mapGrantorDesc' is useful, as it
can help catching other misuses of the index. It must
be the caller's responsibility to ensure that the
grantor index is valid.

Bug: 181269159
Test: adb shell logcat | grep FMQ
      play audio, verify that no error log messages appear
Change-Id: Ib3375be9dca7a14ec840ff723ce26daccc17efad