Add a fuzzer for libfmq
All "random" numbers come from a seed that is provided by the fuzzing
framework.
All excercises are done for AIDL and the default versions of the message
queue.
The fuzzer starts up a writer that creates a new message queue with a
random number of elements. It spawns reader threads passing the MQ
descriptor to each of them. One reader for sync queues, and multiple
readers for unsync queues.
The writer writes a random number of times while the readers reads a
random number of times.

Test: m fmq_fuzzer
Test: adb root && adb sync data
Test: adb shell /data/fuzz/$(get_build_var TARGET_ARCH)/${FUZZER_NAME}/${FUZZER_NAME}"
Bug: 165621851

Change-Id: I529b2ce7fd174d6cb893403a7a345df4434c0ec6