Avoid getIntegralInRange with invalid range
Range of 1 to 0 will abort, and getQuantumCount can return 0 for a queue
with size of 0.
Range of 0 to 0 will not consume any bytes from the FDP, so we can't
loop around that case expecting the FDP to run out.
Changing to range 0 to quantumCount + 1 give some extra coverage and
avoids these issues.

Test: fmq_fuzzer with reproducer in bug
Bug: 252302860
Change-Id: Ib95af7f80c795761461ab2904737323162ba4129